Use sendfile() for gzipped files as well
If 'file.txt.gz' exists when requesting 'file.txt', and if it is newer
than 'file.txt', Lwan will serve this file using sendfile().  This
second stat() call is cached, so the overhead is minimal.

Right now, this only works if Accept-Encoding contains gzip as a
supported encoding (unlike the 'deflate' which was the only supported
encoding up to now). Zlib could be used to obtain the header size and
serve from a certain offset to support the deflate encoding as well.